feat(maven): Support GCP artifact registry (#1155)
Introduce support for GCP Artifact Registry as our preferred remote registry. The Nexus Repository Manager is only supported in fallback mode until a project contains the artifact registry maven wagon extension.

For the artifact registry to work the following must all be true

The project is configured to build with commonpom:v4.0.0 or has been reconfigured to use artifact-registry-releases in distribution management
The project contains a .mvn/extensions.xml with the artifact registry extension
The service-account-key is passed to the action